first commit [FIX] web_editor: fix the offset of form field
The "Form" template uses col-X classes, unlike most snippets that use
col-lg-X classes. This discrepancy was not accounted for when adding an
offset to a form field, causing the form field width to default to 1
when attempting to add the offset.
This fix ensures that when no col-lg-X class is present, the form field
defaults to full-width (col-12) rather than the minimum width.

Steps to reproduce :

- Enter edit mode.
- Insert a "Form" snippet.
- Select a field of the form.
- Add a left margin by dragging the left arrow.
- Bug: The field's width is incorrectly reduced to 1 column.

task-4370763

closes odoo/odoo#193008

Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>